implement async logic and call services in console application

+++ src/main/java/com/martinacat/berriesfinder/view/TotalPrice.java
@@ -1,13 +1,11 @@
 package com.martinacat.berriesfinder.view;
 
-import java.math.BigDecimal;
-
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
 public class TotalPrice {
-    private final BigDecimal gross;
-    private final BigDecimal vat;
+    private final Double gross;
+    private final Double vat;
 }
